[Feature #160245824] Creates Place Order endpoint
