+++ app/__init__.py
@@ -10,11 +10,11 @@ from app.resources.order_resources import SpecificOrder
 from app.resources.order_resources import Views
 
 #challenge 3 imports
-from app.resources.v2.resources.views import OrderResourcesV2
-from app.resources.v2.resources.views import ViewsV2
-from app.resources.v2.resources.views import SpecificOrderV2
+from app.resources.v2.resources.users.orders import OrderResourcesV2
+# from app.resources.v2.resources.users. import ViewsV2
+# from app.resources.v2.resources.views import SpecificOrderV2
 from app.resources.v2.resources.users.user import UserResource
-
+from app.resources.v2.resources.users.user import UserLogin
 
 
 def create_app(config_name):
@@ -39,11 +39,12 @@ def create_app(config_name):
     api.add_resource(SpecificOrder,'/api/v1/orders/<int:order_id>' )
 
     #register endpoints for challenge 3
-    api.add_resource(ViewsV2,'/')
+    # api.add_resource(ViewsV2,'/')
     api.add_resource(OrderResourcesV2, '/api/v2/orders/')
-    api.add_resource(SpecificOrderV2,'/api/v2/orders/')
+    # api.add_resource(SpecificOrderV2,'/api/v2/orders/')
 
     #register endpoiny for challenge 3 user
     api.add_resource(UserResource,'/api/v2/users/')
+    api.add_resource(UserLogin, '/api/v2/users/')
 
     return app
